solve incorrect GL textures being used for seemingly random art tiles
MD4 hashing the attributes identifying an ART tile was giving hash
collisions, and there was nothing further to identify the specific
texture to use. So, md4 is out (and lzf as nothing uses it in BUILD
like I'd intended), and a big lump of memory for a unique identifier
is in. CRC32 of the ident structure helps to quickly eliminate
candidates, but a full memcmp determines which PTMHash to choose.