normalise feature macros
Voxels are now standard equipment and no longer need SUPERBUILD defined.
POLYMOST is now USE_POLYMOST and must be defined with a nonzero value
to activate it. USE_OPENGL also must be defined with a nonzero value
(1 or 2, though it's not validated -- yet) to be activated.

Lastly, NOASM is now USE_ASM (i.e. logically negated) and will only be
activated by default on X86_32 architectures.