fix a sprite vs maskwall rendering order anomaly
Specifically, the presence of sprite A in the layout below causes
sprites B and C to become invisible behind masked wall M the farther
from the player that A is positioned. Turns out we hadn't been doing
the sprite/bunch occlusion test against the correct bunch arrays while
in Polymost mode since forever.

+--------------------------+
|                   A      |
|            /\            |
|       B   / M            |
|          / M             |
|     C   / M              |
|        / M               |
|       / M                |
|       \M                 |
|            /P\           |
+--------------------------+