make permanent rotatesprite 'just work' in GL mode
The game still needs to make certain concessions, like handling the new
special 'numpages' value of 127 which indicates the framebuffer is
invalid from one page to the next. In any case, this change means I
don't need to short-circuit much code on the game side like before to
explicitly force redraw of status bars and borders every frame.

Because of the long persistence of permanent sprites, it also became
necessary to add empty permfifo slot reaping to rotatesprite(). When
the permfifo fills completely it also starts stealing from the tail if
no space at all exists. The old behaviour under DOS would wrap the fifo,
but it was improbable it would ever fill to begin with.

I doubled the maximum length of permfifo to account for lots of sprites
now being tracked, like with the borders of small viewports at large
screen resolutions. Who's likely to make use of that these days besides
a few nostalgic people? *shrug* Memory is cheap now... It may need to
increase again for 4K resolutions, but I haven't done the maths to know.

To accommodate scaling of status bars while still allowing new permanent
rotatesprite() calls to evict sprites that they overlap, I removed the
z==65536 check, making it match with sprites of the same zoom value.
This is fine for simple status bar patch-and-update uses but a change
of scale ought to flush all permanents and refresh.

To that end, in KenBuild, the statusbar and screen border are completely
redrawn on screen resize, because the top and bottom sections were
leaving behind permanent sprites due to the simplicity of the overlap
tests in rotatesprite().