fix macOS 8-bit fullscreen rendering, changing letter/pillarboxing
On macOS 11 with FULLSCREEN_DESKTOP mode, without the HIGHDPI flag the
OpenGL viewport gets initialised in the bottom left corner. With the
HIGHDPI flag, the GL viewport stretches across the full screen, noting
SDL_GL_GetDrawableSize() just returns the dimensions that we passed to
SDL_CreateWindow(). The actual fullscreen resolution is obtained by
querying the display mode related to the window. We correct for the
mismatch in aspect ratio between the physical resolution and the logical
surface by skewing the quad that gets drawn. Since we're going to all
that effort, on Windows we now also just use a full desktop-sized GL
window and render the skewed quad.