adapt to C99
The ambition was once to accommodate platforms that didn't have a
real C standard library, the particular one in my mind at the time was
PalmOS. That never happened, and with C99 (and earlier) having
standardised nearly everything I had B-prefixed to be later shimmed,
there's no reason anymore to have compat.h doing so much.

GCC 4.5 was practically C99 compliant, and MSVC from 2015 onwards is
broadly C89-99 and C++98 compliant, so the plan is now: set the barrier
to entry to C99 and C++98, and adapt to those specifications.

There will now be a lot of signedness and precision loss warnings to
fix in the following commits.