massive GL function pointer casting endeavour
Function pointer types not in gl.h or glext.h blagged from glcorearb.h.

On that note, for GL2 and GL3 we use the same headers. Using glcorearb.h
loses us a lot of extensions, and it claims you must not mix it with
glext.h. So although we ask for a core profile in GL3, we compile as
though we're in a compatibility world. Nonetheless, there's now ifdefs
surrounding the GL_EXT_texture_filter_anisotropic and
GL_NV_multisample_filter_hint uses in case someone does indeed choose
glcorearb.h (which is commented out in glbuild_priv.h).

And while I was here, I amended the debug extension to cover
GL_ARB_debug_output.