finesse mouse sensitivity in the editor
msens is only used in build.c so define it there, and why not remove
the use of floating point for sensitivity scaling just because, so the
config read/writer ought be able to handle 16:16 fixed-point directly
and read/write it as a fractional value.

Also observed that the already-scaled surpluses were being fed back in
pre-scale, which caused the cursor to drift when the mouse was sitting
stationary at settings > 1.0.