fix fullscreen behaviour on linux and macos
On linux the glbuild frame would be displayed at weird size in the
corner of the display, apparently because the drawable size only becomes
known once the window is shown.

On macOS 15 it doesn't seem necessary to request high-DPI since the GL
viewport will be 0.25x the size of the context's drawable size. The
fullscreen modes on my 3024x1964 retina display all seem to behave as
they're expected to.