review key repeating again
In build.c's original keyhandler(), the single-interrupt keys were
repeat-suppressed and double-interrupt keys (e.g. arrow keys) were not.
My change to make all keys non-repeating in b7633cfcd40 means the editor
file selector and tile chooser would not repeat the arrow keys when
navigating.

This change isn't exactly like the DOS behaviour in that keystatus[]
remains non-repeating and reflects the pressed-or-not state (unless
the game modifies the key), but repeating is added to keyfifo[] (and
releases also reported), and build.c is adapted to read from the fifo.
And any of the keys in those cases will repeat whether they used to or
not.